Moves a lot of duplicated dataset management code to shared util file
